+++ lib/flutter_input_library.dart
@@ -4,5 +4,5 @@
 /// A library for creating input fields in Flutter.
 library flutter_input_library;
 
-export 'src/inputs/inputs.dart';
-export 'src/utils/utils.dart';
+export "src/inputs/inputs.dart";
+export "src/utils/utils.dart";