+++ lib/src/inputs/carousel/carousel.dart
@@ -2,9 +2,9 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-import 'package:flutter/material.dart';
+import "package:flutter/material.dart";
 
-import 'package:flutter_input_library/src/inputs/carousel/carousel_form.dart';
+import "package:flutter_input_library/src/inputs/carousel/carousel_form.dart";
 
 class FlutterFormInputCarousel extends StatelessWidget {
   const FlutterFormInputCarousel({