+++ lib/src/inputs/phone/phone.dart
@@ -3,9 +3,9 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 // ignore: depend_on_referenced_packages
-import 'package:collection/collection.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_input_library/flutter_input_library.dart';
+import "package:collection/collection.dart";
+import "package:flutter/material.dart";
+import "package:flutter_input_library/flutter_input_library.dart";
 
 class FlutterFormInputPhone extends StatefulWidget {
   const FlutterFormInputPhone({
@@ -20,7 +20,7 @@ class FlutterFormInputPhone extends StatefulWidget {
     this.numberFieldStyle,
     this.dialCodeSelectorStyle,
     this.enabled = true,
-    this.priorityCountries = const ['NL', 'BE', 'LU'],
+    this.priorityCountries = const ["NL", "BE", "LU"],
     this.textAlignVertical = TextAlignVertical.top,
     this.dialCodeSelectorPadding = const EdgeInsets.only(top: 6),
   });
@@ -82,7 +82,7 @@ class _FlutterFormInputPhoneState extends State<FlutterFormInputPhone> {
   Widget build(BuildContext context) {
     var inputDecoration = widget.decoration ??
         InputDecoration(
-          label: widget.label ?? const Text('Phone number'),
+          label: widget.label ?? const Text("Phone number"),
         );
 
     return FlutterFormInputPlainText(
@@ -122,7 +122,7 @@ class _FlutterFormInputPhoneState extends State<FlutterFormInputPhone> {
                       const SizedBox(
                         width: 4,
                       ),
-                      Text('+${country.dialCode}'),
+                      Text("+${country.dialCode}"),
                     ],
                   ),
                 ),