+++ lib/src/inputs/slider/slider_field.dart
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-import 'package:flutter/material.dart';
+import "package:flutter/material.dart";
 
 /// Creates a slider with the given input parameters
 class SliderFormField extends FormField<double> {