+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.6.0
 * Added style parameter to `FlutterFormInputMultiLine`
+* Upgrade flutter_iconica_analysis to 7.0.0
 
 ## 3.5.0
 * Added `selectableTimePredicate` to the `DateTimeInputField` constructor